Updated files
Last commit consisted of outdated files.

+++ delete-coinflip.php
@@ -1,5 +1,5 @@
 <?php
-if (isset($_POST["id"]) && isset($_POST["sessionId"])) {
+if (isset($_POST["id"]) && ctype_digit($_POST["id"]) && isset($_POST["sessionId"])) {
 	$db = mysqli_connect();
 	$db->select_db("coinflip");
 	
@@ -9,12 +9,20 @@ if (isset($_POST["id"]) && isset($_POST["sessionId"])) {
 	$result = $stmt->get_result()->fetch_assoc();
 	
 	if ($result) {
-		$stmt = $db->prepare("DELETE FROM coinflips WHERE id=? AND user=?;");
-		$stmt->bind_param("ii", $_POST["id"], $result["user"]);
-		$stmt->execute();
-		if ($db->affected_rows > 0) {
-			echo "ok";
-			exit();
+		$cf = $db->query("SELECT * FROM coinflips WHERE id=".$_POST["id"])->fetch_assoc();
+		if ($cf) {
+			$stmt = $db->prepare("DELETE FROM coinflips WHERE id=? AND user=?;");
+			$stmt->bind_param("ii", $_POST["id"], $result["user"]);
+			$stmt->execute();
+			
+			$stmt = $db->prepare("UPDATE users SET balance=balance+? WHERE id=?;");
+			$stmt->bind_param("ii", $cf["bet"], $result["user"]);
+			$stmt->execute();
+			
+			if ($db->affected_rows > 0) {
+				echo "ok";
+				exit();
+			}
 		}
 	}
 }
